requirements.txt
python-telegram-bot[job-queue]>=21.0
aiosqlite>=0.20.0
google-genai>=1.0.0
matplotlib>=3.9.0
python-dotenv>=1.0.0
